mysite/requirements.txt
Django>=4.0
psycopg2-binary>=2.9
django-cors-headers>=4.0
# Django y utilidades core
Django>=5.1.1
python-decouple>=3.8

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# CORS headers para conectar con frontend
django-cors-headers>=4.3.0

# Filtros y paginación avanzada
django-filter>=23.0

# Base de datos y migraciones
psycopg2-binary>=2.9.9  # Para PostgreSQL (opcional)

# Manejo de imágenes
Pillow>=10.0.0

# MQTT para IoT/SCADA
paho-mqtt>=1.6.1

# Utilidades adicionales
python-dateutil>=2.8.2
pytz>=2023.3

# Para desarrollo
django-extensions>=3.2.3

# APIs y documentación
drf-spectacular>=0.26.0  # Para generar documentación OpenAPI/Swagger
